feat: RemoteAPI, save index->index.old, working on Remote API for Session management via Websockets
