+++ build_local.sh
@@ -1,29 +1,29 @@
 # npm run clean
 # npm install
 
-set -e
+# set -e
 # Check Emscripten installation
-if ! [[ $(which emcmake) ]]; then
-    echo "emcmake not found on local environment !"
-    echo "Please install emscripten: https://emscripten.org/docs/getting_started/downloads.html"
-    exit 1
-else
-    echo "emcmake found on local environment !"
-    echo $(which emcmake)
-fi
+# if ! [[ $(which emcmake) ]]; then
+#     echo "emcmake not found on local environment !"
+#     echo "Please install emscripten: https://emscripten.org/docs/getting_started/downloads.html"
+#     exit 1
+# else
+#     echo "emcmake found on local environment !"
+#     echo $(which emcmake)
+# fi
 
 # Begin build
 
-echo "Start Build Locally"
+# echo "Start Build Locally"
 
-cd ./zenoh-wasm 
+# cd ./zenoh-wasm 
 
-bash build-commands.sh 
+# bash build-commands.sh 
 
-cd ../
+# cd ../
 
-rm -rf ./src/wasm
+# rm -rf ./src/wasm
 
-cp -r ./zenoh-wasm/build.emscripten ./src/wasm
+# cp -r ./zenoh-wasm/build.emscripten ./src/wasm
 
 npm run build:ts
