+++ package.json
@@ -16,7 +16,9 @@
     "src"
   ],
   "dependencies": {
+    "channel-ts": "^0.1.2",
     "cpx": "^1.5.0",
+    "fp-ts": "^2.16.5",
     "object-inspect": "^1.13.1",
     "request": "^2.79.0",
     "text-encoding": "^0.7.0",
@@ -47,7 +49,6 @@
     "build:copy_artifacts": "npm run build:copy_zenoh_js && npm run build:copy_wasm",
     "build:copy_zenoh_js": "cpx -v \"src/wasm/**/*.{wasm,js,mjs,}\" \"esm/wasm\" && cpx -v \"src/wasm/**/*.{wasm,js,mjs,}\" \"cjs/wasm\"",
     "build:copy_wasm": "cpx -v \"src/**/*.wasm\" \"./cjs\" && cpx -v \"src/**/*.wasm\" \"./esm\"",
-   
     "//   PACKAGE       //": "",
     "webpack": "webpack",
     "example": "cd examples && npm run start"
